# Optimize 'json_parse_string' using SSE2.